Fix integer overflow warning
As the evaluation of the righthand side of the expression is done before
assignment to the left-side unsigned long variable, the multiplication of the
three integer values (result 60000) exceeds the upper bound for signed integer
values.

Suffixing the first right-hand value with `L` tells the compiler that the value
is of type `long`, which causes the expression to be evaluated as the
multiplication of `long` values instead of `int` values.